backend/requirements.txt
requests
beautifulsoup4
cerebras-cloud-sdk
flask